Add the cement kiln environment
A 3000 t/day dry-process rotary kiln with preheater and precalciner, modelled
as 16 axial slices each carrying solid temperature, refractory temperature,
calcination extent and free lime, with the gas solved quasi-steady by a
counter-current sweep (justified: gas crosses the kiln in 8 s against 25 min
for the solid).

This brings the suite's TRANSPORT DELAY problem, and the delay is emergent --
there is no delay parameter anywhere. Solid states are advected down the kiln,
so half the free-lime response to a fuel change takes ~22 min, one full
residence time, and 90 % takes ~80 min. Kiln speed moves the delay itself:
2.4 rpm gives 31 min residence and 0.8 % lime, 4.2 rpm gives 18 min and 2.4 %,
at nearly unchanged burning-zone temperature. That makes the second input
qualitatively different from the first rather than another knob on the same
response. 64 states are dynamic behind 8 measurements, which is roughly how
blind real kiln operation is.

An energy audit drove the derivation and caught the main error: the kiln was
being fed RAW meal when a precalciner kiln receives meal that has already shed
92 % of its CO2, overstating its thermal load by nearly 50 % (53.8 vs 36.6
kg/s). Four more surfaced the same way -- Arrhenius pre-exponentials implying
calcination in 0.25 s and clinkerisation in 0.45 s (violently stiff, NaN); a
point-source flame dumping the whole firing rate into one 3.75 m slice, which
exhausted the gas and left the kiln colder than its own feed; gas emissivity of
0.8 rather than the 0.25 of CO2/H2O bands; and transfer areas that ignored the
covered arc, which is a major regenerative route at 10 % fill. Energy now
closes to 0.00 % on the converged profile, and the burner duty agrees with the
published 40 % split by an independent bottom-up route (43.9 vs 44.4 MW).

The MPC is gradient-free by necessity, not preference. Free lime depends on
temperature through a 280 kJ/mol Arrhenius term and that temperature is
advected, so the tangent system grows about two orders of magnitude per step:
reverse-mode gradients overflow to NaN after ~8 steps (1.6e-3 over five, 8.3e3
over eight) while finite differences on the identical objective stay clean.
GradientMPC silently zeroes NaN gradients, so it sat at its initial sequence
and scored exactly what a zero-action constant scores. A new SamplingMPC (CEM)
samples instead and recovers the expected margin: 193-203 against 163-180 for
the cascade PID and 37 for the best constant action.

Cascade is the right PID structure and the reason is structural -- fuel on the
burning-zone pyrometer (fast, no delay) with free lime only trimming that
setpoint slowly, and kiln speed following feed to keep bed depth and therefore
the delay from wandering. Gains reflect it: strongly proportional outer loop
with weak integral, because integral action on a half-hour-old measurement
oscillates at the delay period.

22 conformance tests pass on first registration, plus 32 physics tests.
Throughput 0.65 M steps/s -- the slowest environment here, and honestly so.
Suite: 950 passed, 0 failed.

+++ src/target_gym/boiler_drum/env_jax.py
@@ -115,16 +115,16 @@ class BoilerDrum(environment.Environment[BoilerDrumState, BoilerDrumParams]):
 
         state = BoilerDrumState(
             time=0,
-            pressure=jnp.asarray(params.p_nominal, dtype=jnp.float32),
+            pressure=jnp.asarray(params.p_nominal),
             V_wt=V_wt,
             m_sr=m_sr,
             m_sd=m_sd,
-            level=jnp.asarray(level0, dtype=jnp.float32),
+            level=jnp.asarray(level0),
             q_steam=q_steam,
-            Q_fuel=jnp.asarray(Q, dtype=jnp.float32),
-            q_feed=jnp.asarray(params.q_steam_nominal, dtype=jnp.float32),
+            Q_fuel=jnp.asarray(Q),
+            q_feed=jnp.asarray(params.q_steam_nominal),
             target_pressure=target_pressure,
-            level_ref=jnp.asarray(level_ref, dtype=jnp.float32),
+            level_ref=jnp.asarray(level_ref),
         )
         return self.get_obs(state), state
